[IR] Refactor IR classes to enforce copy-on-write machanism
This PR refactors the IR classes to use `dataclass(frozen=True)` to
prevent we change the IR nodes once it is created. Whenever we want to
change the IR nodes, we need to create a new one with the modification,
and reuse any sub-components that are not changed (i.e., copy-on-write
machanism).