[Layout] Add the unified representation of layout system
Previously, we use a layout-tree to represent the layouts.

This PR implements a unified underlying representation of tilus' layout
system for register tensors.

```python
class RegisterLayout:
    shape: list[int]
    mode_shape: list[int]
    spatial_modes: list[int]
    local_modes: list[int]
```

The mode_shape is a fine-grained shape, which may split each dimension
in `shape` into multiple sub-dimensions.

The `spatial` and `local` dims are the dimensions for spatial and local.
Here spatial indicates different workers and local indicate different
entries for the same worker.

PS: this PR also add a cProfile utility to profile a python function.